backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
pytest-flask==1.3.0
python-dotenv==1.0.1